+++ library/login.js
@@ -1,0 +1,4 @@
+const { login_client } = require("./loginFunction.js");
+
+login_client(JSON.parse(process.argv[2]));
+console.log("AGIDs:" + process.argv[2]);